requirements.txt
Django==1.9.2
Pillow==3.1.1
PyJWT==1.4.0
argparse==1.4.0
defusedxml==0.4.1
django-ckeditor==4.5.1
django-endless-pagination==2.0
django-registration==2.0.4
django-stdimage==2.3.2
oauthlib==1.0.3
progressbar2==3.6.2
psycopg2==2.6.1
python-social-auth==0.2.14
python3-openid==3.0.9
requests==2.3.0
requests-oauthlib==0.6.1
six==1.10.0
steamapi==0.1
